+++ src/main/java/uz/md/shopapp/service/impl/InstitutionServiceImpl.java
@@ -13,6 +13,7 @@ import uz.md.shopapp.dtos.institution.InstitutionDTO;
 import uz.md.shopapp.dtos.institution.InstitutionEditDTO;
 import uz.md.shopapp.dtos.institution.InstitutionInfoDTO;
 import uz.md.shopapp.exceptions.AlreadyExistsException;
+import uz.md.shopapp.exceptions.BadRequestException;
 import uz.md.shopapp.exceptions.NotAllowedException;
 import uz.md.shopapp.exceptions.NotFoundException;
 import uz.md.shopapp.mapper.InstitutionMapper;
@@ -20,10 +21,11 @@ import uz.md.shopapp.repository.InstitutionRepository;
 import uz.md.shopapp.repository.UserRepository;
 import uz.md.shopapp.service.contract.FilesStorageService;
 import uz.md.shopapp.service.contract.InstitutionService;
-import uz.md.shopapp.utils.AppConstants;
 
 import java.nio.file.Path;
 import java.util.List;
+
+import static uz.md.shopapp.utils.MessageConstants.*;
 
 @Service
 public class InstitutionServiceImpl implements InstitutionService {
@@ -31,7 +33,7 @@ public class InstitutionServiceImpl implements InstitutionService {
     @Value("${app.images.institutions.root.path}")
     private String institutionsPath;
 
-    private Path institutionsImagesRoot ;
+    private Path institutionsImagesRoot;
 
     @PostConstruct
     public void init() {
@@ -55,6 +57,12 @@ public class InstitutionServiceImpl implements InstitutionService {
 
     @Override
     public ApiResult<InstitutionDTO> add(InstitutionAddDTO dto) {
+
+        if (dto == null || dto.getNameUz() == null || dto.getNameRu() == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
 
         if (institutionRepository.existsByNameUzOrNameRu(dto.getNameUz(), dto.getNameRu()))
             throw AlreadyExistsException.builder()
@@ -86,23 +94,41 @@ public class InstitutionServiceImpl implements InstitutionService {
 
     @Override
     public ApiResult<InstitutionDTO> findById(Long id) {
+        if (id == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
+
         return ApiResult.successResponse(institutionMapper
                 .toDTO(institutionRepository
                         .findById(id)
                         .orElseThrow(() -> NotFoundException.builder()
-                                .messageUz("INSTITUTION_NOT_FOUND")
-                                .messageRu("")
+                                .messageUz(INSTITUTION_NOT_FOUND_UZ)
+                                .messageRu(INSTITUTION_NOT_FOUND_RU)
                                 .build())));
     }
 
     @Override
     public ApiResult<InstitutionDTO> edit(InstitutionEditDTO editDTO) {
 
+        if (editDTO == null
+                || editDTO.getNameUz() == null
+                || editDTO.getId() == null
+                || editDTO.getNameRu() == null
+                || editDTO.getLocation() == null
+                || editDTO.getInstitutionTypeId() == null
+                || editDTO.getManagerId() == null)
+            throw BadRequestException.builder()
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .build();
+
         Institution editing = institutionRepository
                 .findById(editDTO.getId())
                 .orElseThrow(() -> NotFoundException.builder()
-                        .messageUz("INSTITUTION_NOT_FOUND")
-                        .messageRu("")
+                        .messageUz(INSTITUTION_NOT_FOUND_UZ)
+                        .messageRu(INSTITUTION_NOT_FOUND_RU)
                         .build());
 
         if (institutionRepository.existsByNameUzOrNameRuAndIdIsNot(editDTO.getNameUz(), editDTO.getNameRu(), editing.getId()))
@@ -132,18 +158,33 @@ public class InstitutionServiceImpl implements InstitutionService {
 
     @Override
     public ApiResult<List<InstitutionInfoDTO>> getAllByTypeId(Long typeId) {
+        if (typeId == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
         return ApiResult.successResponse(institutionRepository
                 .findAllForInfoByTypeId(typeId));
     }
 
     @Override
     public ApiResult<List<InstitutionInfoDTO>> getAllByManagerId(Long managerId) {
+        if (managerId == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
         return ApiResult.successResponse(institutionRepository
                 .findAllForInfoByManagerId(managerId));
     }
 
     @Override
     public ApiResult<List<InstitutionInfoDTO>> getAllForInfoByPage(String page) {
+        if (page == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
         int[] paged = {Integer.parseInt(page.split("-")[0]),
                 Integer.parseInt(page.split("-")[1])};
         return ApiResult.successResponse(institutionRepository
@@ -153,11 +194,16 @@ public class InstitutionServiceImpl implements InstitutionService {
 
     @Override
     public ApiResult<Void> setImage(Long institutionId, MultipartFile image) {
+        if (institutionId == null || image == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
         Institution institution = institutionRepository
                 .findById(institutionId)
                 .orElseThrow(() -> NotFoundException.builder()
-                        .messageUz("INSTITUTION NOT FOUND")
-                        .messageRu("")
+                        .messageUz(INSTITUTION_NOT_FOUND_UZ)
+                        .messageRu(INSTITUTION_NOT_FOUND_RU)
                         .build());
         filesStorageService.save(image, institutionsImagesRoot);
         institution.setImageUrl(institutionsImagesRoot.toUri() + image.getOriginalFilename());
@@ -167,6 +213,11 @@ public class InstitutionServiceImpl implements InstitutionService {
 
     @Override
     public ApiResult<Void> delete(Long id) {
+        if (id == null)
+            throw BadRequestException.builder()
+                    .messageUz(ERROR_IN_REQUEST_UZ)
+                    .messageRu(ERROR_IN_REQUEST_RU)
+                    .build();
         institutionRepository.deleteById(id);
         return ApiResult.successResponse();
     }
